refresh problem

+++ backend/src/friendRequest/friendRequest.gateway.ts
@@ -32,7 +32,6 @@ export class FriendRequestGateway {
 			(await receiver).friendRequests.push(await pushFriendRequest);
 			await this.usersService.save(sender);
 			await this.usersService.save(receiver);
-			this.dmService.dmUsers.find(user => user.name === request.sender).socket.emit('send_friendRequest');
 			this.dmService.dmUsers.find(user => user.name === request.receiver).socket.emit('send_friendRequest');
 		}
 	}
@@ -42,19 +41,21 @@ export class FriendRequestGateway {
 		const sender = this.usersService.getByNameWithRelations(request.sender);
 		const receiver = this.usersService.getByNameWithRelations(request.receiver);
 		const toDelete = (await this.friendRequestService.getFriendRequestByUserId((await sender).id, (await receiver).id));
-		let i = (await sender).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
-		(await sender).friendRequests.splice(i, 1);
-		i = (await receiver).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
-		(await receiver).friendRequests.splice(i, 1);
-		await this.friendRequestService.remove(toDelete);
-		const senderFriend: IFriend = { user: await sender, friendOf: await receiver };
-		const receiverFriend : IFriend = { user: await receiver, friendOf: await sender };
-		const pushSenderFriend = this.friendRequestService.createNewFriend(senderFriend);
-		const pushReceiverFriend = this.friendRequestService.createNewFriend(receiverFriend);
-		(await sender).friends.push(await pushReceiverFriend);
-		(await receiver).friends.push(await pushSenderFriend);
-		await this.usersService.save(sender);
-		await this.usersService.save(receiver);
+		if (toDelete !== undefined) {
+			let i = (await sender).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
+			(await sender).friendRequests.splice(i, 1);
+			i = (await receiver).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
+			(await receiver).friendRequests.splice(i, 1);
+			await this.friendRequestService.remove(toDelete);
+			const senderFriend: IFriend = { user: await sender, friendOf: await receiver };
+			const receiverFriend : IFriend = { user: await receiver, friendOf: await sender };
+			const pushSenderFriend = this.friendRequestService.createNewFriend(senderFriend);
+			const pushReceiverFriend = this.friendRequestService.createNewFriend(receiverFriend);
+			(await sender).friends.push(await pushReceiverFriend);
+			(await receiver).friends.push(await pushSenderFriend);
+			await this.usersService.save(sender);
+			await this.usersService.save(receiver);
+		}
 		this.dmService.dmUsers.find(user => user.name === request.sender).socket.emit('accept_friendRequest');
 		this.dmService.dmUsers.find(user => user.name === request.receiver).socket.emit('accept_friendRequest');
 	}
@@ -64,13 +65,15 @@ export class FriendRequestGateway {
 		const sender = this.usersService.getByNameWithRelations(request.sender);
 		const receiver = this.usersService.getByNameWithRelations(request.receiver);
 		const toDelete = (await this.friendRequestService.getFriendRequestByUserId((await sender).id, (await receiver).id));
-		let i = (await sender).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
-		(await sender).friendRequests.splice(i, 1);
-		i = (await receiver).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
-		(await receiver).friendRequests.splice(i, 1);
-		await this.friendRequestService.remove(toDelete);
-		await this.usersService.save(sender);
-		await this.usersService.save(receiver);
+		if (toDelete !== undefined) {
+			let i = (await sender).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
+			(await sender).friendRequests.splice(i, 1);
+			i = (await receiver).friendRequests.findIndex(friendRequest => friendRequest.id === toDelete.id);
+			(await receiver).friendRequests.splice(i, 1);
+			await this.friendRequestService.remove(toDelete);
+			await this.usersService.save(sender);
+			await this.usersService.save(receiver);
+		}
 		this.dmService.dmUsers.find(user => user.name === request.sender).socket.emit('refuse_friendRequest');
 		this.dmService.dmUsers.find(user => user.name === request.receiver).socket.emit('refuse_friendRequest');
 	}
